+++ doc/version_history.md
@@ -1,5 +1,14 @@
 # Version Histroy
 
+0.2.4
+
+- Add the **telemetry_stable_time** to `parameters_power.yaml`.
+- Update the **ConfigPower** to have the expected timing in power on/off process.
+- Improve the logic of breakers in **MockPlant** to have the more realistic behavior.
+- Update the `SubPowerSystem.transition_state()` with the appropriate state machine that tracks the power on/off process based on the timing and others.
+- Improve the **PowerSystem** to rely on **SubPowerSystem** to do the state transition.
+- Add the `disableConcurrentBuilds()` to the **Jenkinsfile**.
+
 0.2.3
 
 - Add the `mock_power_system.rs`.